backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-community
chromadb
pydantic
apscheduler
python-dotenv
httpx
numpy
scikit-learn
sentence-transformers
torch
transformers
accelerate
yfinance
networkx
librosa
soundfile
openai-whisper
